[lint] Use eslint-config-universe (#10)
Why: This is the linter configuration we use in our other repos

How: Changed the ESLint config file to reference universe/node. We can add universe/web if we want to target web in the future but for now our intent is to run this library in Node CLIs.

Added a Prettier configuration file as well.

Test Plan: Ran `yarn lint` and verified it ran successfully. It printed a lot of suggested that I did not apply in this commit in order to keep them separate.